Moved header into components folder
